Split `access.go` into dedicated files
