+++ test/e2e/shoot/internal/access_csr.go
@@ -20,6 +20,12 @@ import (
 	"crypto/x509/pkix"
 	"time"
 
+	"github.com/gardener/gardener/pkg/client/kubernetes"
+	"github.com/gardener/gardener/pkg/controllerutils"
+	"github.com/gardener/gardener/pkg/utils"
+	"github.com/gardener/gardener/pkg/utils/flow"
+	secretutils "github.com/gardener/gardener/pkg/utils/secrets"
+
 	certificatesv1 "k8s.io/api/certificates/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
@@ -29,44 +35,7 @@ import (
 	csrutil "k8s.io/client-go/util/certificate/csr"
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	authenticationv1alpha1 "github.com/gardener/gardener/pkg/apis/authentication/v1alpha1"
-	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
-	gardenversionedcoreclientset "github.com/gardener/gardener/pkg/client/core/clientset/versioned"
-	"github.com/gardener/gardener/pkg/client/kubernetes"
-	"github.com/gardener/gardener/pkg/controllerutils"
-	"github.com/gardener/gardener/pkg/utils"
-	"github.com/gardener/gardener/pkg/utils/flow"
-	gutil "github.com/gardener/gardener/pkg/utils/gardener"
-	secretutils "github.com/gardener/gardener/pkg/utils/secrets"
 )
-
-// CreateShootClientFromStaticTokenKubeconfig retrieves the static token kubeconfig secret and creates a shoot client.
-func CreateShootClientFromStaticTokenKubeconfig(ctx context.Context, gardenClient kubernetes.Interface, shoot *gardencorev1beta1.Shoot) (kubernetes.Interface, error) {
-	return kubernetes.NewClientFromSecret(ctx, gardenClient.Client(), shoot.Namespace, gutil.ComputeShootProjectSecretName(shoot.Name, "kubeconfig"),
-		kubernetes.WithDisabledCachedClient(),
-	)
-}
-
-// CreateShootClientFromAdminKubeconfig requests an admin kubeconfig and creates a shoot client.
-func CreateShootClientFromAdminKubeconfig(ctx context.Context, gardenClient kubernetes.Interface, shoot *gardencorev1beta1.Shoot) (kubernetes.Interface, error) {
-	versionedClient, err := gardenversionedcoreclientset.NewForConfig(gardenClient.RESTConfig())
-	if err != nil {
-		return nil, err
-	}
-
-	adminKubeconfigRequest := &authenticationv1alpha1.AdminKubeconfigRequest{
-		Spec: authenticationv1alpha1.AdminKubeconfigRequestSpec{
-			ExpirationSeconds: pointer.Int64(3600),
-		},
-	}
-	adminKubeconfig, err := versionedClient.CoreV1beta1().Shoots(shoot.GetNamespace()).CreateAdminKubeconfigRequest(ctx, shoot.GetName(), adminKubeconfigRequest, metav1.CreateOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	return kubernetes.NewClientFromBytes(adminKubeconfig.Status.Kubeconfig, kubernetes.WithDisabledCachedClient())
-}
 
 // labelsE2ETestCSRAccess is the set of labels added to all CSRs and ClusterRoleBindings for easy cleanup.
 var labelsE2ETestCSRAccess = map[string]string{"e2e-test": "csr-access"}
@@ -85,12 +54,19 @@ func CreateShootClientFromCSR(ctx context.Context, shootClient kubernetes.Interf
 		return nil, err
 	}
 
-	reqName, reqUID, err := csrutil.RequestCertificate(shootClient.Kubernetes(), csrData, commonName,
-		certificatesv1.KubeAPIServerClientSignerName, pointer.Duration(3600*time.Second), []certificatesv1.KeyUsage{
+	reqName, reqUID, err := csrutil.RequestCertificate(
+		shootClient.Kubernetes(),
+		csrData,
+		commonName,
+		certificatesv1.KubeAPIServerClientSignerName,
+		pointer.Duration(3600*time.Second),
+		[]certificatesv1.KeyUsage{
 			certificatesv1.UsageDigitalSignature,
 			certificatesv1.UsageKeyEncipherment,
 			certificatesv1.UsageClientAuth,
-		}, privateKey)
+		},
+		privateKey,
+	)
 	if err != nil {
 		return nil, err
 	}
@@ -106,11 +82,9 @@ func CreateShootClientFromCSR(ctx context.Context, shootClient kubernetes.Interf
 		return nil, err
 	}
 
-	clusterRoleBinding := &rbacv1.ClusterRoleBinding{
-		ObjectMeta: metav1.ObjectMeta{Name: commonName, Labels: labelsE2ETestCSRAccess},
-	}
-
+	clusterRoleBinding := &rbacv1.ClusterRoleBinding{ObjectMeta: metav1.ObjectMeta{Name: commonName}}
 	if _, err = controllerutils.GetAndCreateOrMergePatch(ctx, shootClient.Client(), clusterRoleBinding, func() error {
+		clusterRoleBinding.Labels = utils.MergeStringMaps(clusterRoleBinding.Labels, labelsE2ETestCSRAccess)
 		clusterRoleBinding.RoleRef = rbacv1.RoleRef{
 			APIGroup: rbacv1.GroupName,
 			Kind:     "ClusterRole",
